+++ sketch-animated-gif-koral/src/androidTest/kotlin/com/github/panpf/sketch/animated/gif/koral/test/decode/internal/KoralGifDecoderComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.gif.koral.test.decode.internal
+
+class KoralGifDecoderComponentTest {
+    // TODO test
+}