+++ sketch-animated-gif/src/jvmCommonTest/kotlin/com/github/panpf/sketch/animated/gif/jvm/test/decode/internal/GifDecoderComponentJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.gif.jvm.test.decode.internal
+
+class GifDecoderComponentJvmTest {
+    // TODO test
+}