+++ sketch-animated-gif/src/nonJvmCommonTest/kotlin/com/github/panpf/sketch/animated/gif/nonjvm/test/decode/internal/GifDecoderComponentNonJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.gif.nonjvm.test.decode.internal
+
+class GifDecoderComponentNonJvmTest {
+    // TODO test
+}