+++ sketch-animated-heif/src/androidTest/kotlin/com/github/panpf/sketch/animated/heif/test/decode/internal/ImageDecoderAnimatedHeifDecoderComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.heif.test.decode.internal
+
+class ImageDecoderAnimatedHeifDecoderComponentTest {
+    // TODO test
+}