+++ sketch-animated-webp/src/jvmCommonTest/kotlin/com/github/panpf/sketch/animated/webp/jvm/test/decode/internal/AnimatedWebpDecoderComponentJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.webp.jvm.test.decode.internal
+
+class AnimatedWebpDecoderComponentJvmTest {
+    // TODO test
+}